+++ modules/apps/collaboration/reading-time/reading-time-service/src/test/java/com/liferay/reading/time/internal/calculator/ReadingTimeCalculatorImplTest.java
@@ -19,7 +19,7 @@ import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.reading.time.calculator.ReadingTimeCalculator;
-import com.liferay.reading.time.calculator.ReadingTimeCalculatorImpl;
+import com.liferay.reading.time.calculator.impl.ReadingTimeCalculatorImpl;
 
 import java.time.Duration;
